// static_report/src/sdlc/comparison-report-schema.d.ts
export interface ComparisonReportSchema {
// ...
  input?: {
    tables?: {
// ...
      ACTION?: {
        name?: string;
        row_count?: number;
        col_count?: number;
        columns?: {
          symbol?: {
            total?: number;
            non_nulls?: number;
            distinct?: number;
            distribution?: {
              type?: string;
              labels?: string[];
              counts?: number[];
              [k: string]: unknown;
            };
            name?: string;
            profile_duration?: string;
            type?: string;
            schema_type?: string;
            [k: string]: unknown;
          };
          date?: {
            total?: number;
            non_nulls?: number;
            mismatched?: number;
            distinct?: number;
            min?: string;
            max?: string;
            distribution?: {
              type?: string;
              labels?: string[];
              counts?: number[];
              [k: string]: unknown;
            };
            name?: string;
            profile_duration?: string;
            type?: string;
            schema_type?: string;
            [k: string]: unknown;
          };
          dividends?: {
            total?: number;
            non_nulls?: number;
            mismatched?: number;
            distinct?: number;
            min?: number;
            max?: number;
            sum?: number;
            avg?: number;
            p5?: number;
            p25?: number;
            stddev?: number;
            p50?: number;
            p75?: number;
            p95?: number;
            distribution?: {
              type?: string;
              labels?: string[];
              counts?: number[];
              bin_edges?: number[];
              [k: string]: unknown;
            };
            name?: string;
            profile_duration?: string;
            type?: string;
            schema_type?: string;
            [k: string]: unknown;
          };
          splits?: {
            total?: number;
            non_nulls?: number;
            mismatched?: number;
            distinct?: number;
            min?: number;
            max?: number;
            sum?: number;
            avg?: number;
            p5?: number;
            p25?: number;
            stddev?: number;
            p50?: number;
            p75?: number;
            p95?: number;
            distribution?: {
              type?: string;
              labels?: string[];
              counts?: number[];
              bin_edges?: number[];
              [k: string]: unknown;
            };
            name?: string;
            profile_duration?: string;
            type?: string;
            schema_type?: string;
            [k: string]: unknown;
          };
          [k: string]: unknown;
        };
        assertion_results?: {
          tests?: unknown[];
          columns?: {
            [k: string]: unknown;
          };
          [k: string]: unknown;
        };
        description?: string;
        [k: string]: unknown;
      };
// ...
    };
    id?: string;
    created_at?: string;
    datasource?: {
      name?: string;
      type?: string;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}